backend/createHTML/requirements_bak.txt
appdirs==1.4.3
beautifulsoup4==4.9.3
bs4==0.0.1
cycler==0.10.0
distlib==0.3.0
filelock==3.0.12
hitherdither==0.1.6
importlib-metadata==1.6.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.2.0
Pillow==8.1.0
pymodbus==2.3.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.5
six==1.14.0
soupsieve==2.1
virtualenv==20.0.17
zipp==3.1.0
